// src/lib/api/generated/model/exitInterviewsQueueListParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The canonical backend API for COMPASS clients.
 * OpenAPI spec version: 1.0.0
 */

export type ExitInterviewsQueueListParams = {
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
page_size?: number;
q?: string;
status?: string;
academic_year?: string;
revision?: string;
order?: string;
};